refactor: reorganize budgets methods
