+++ src/entities/QuarterBudget/model.ts
@@ -1,41 +1,13 @@
 import Time from 'dayjs'
 import { Model } from 'decentraland-gatsby/dist/entities/Database/model'
-import { snakeCase } from 'lodash'
 import { v1 as uuid } from 'uuid'
 
 import { TransparencyBudget } from '../../clients/DclData'
-import { ProposalGrantCategory } from '../Proposal/types'
 import QuarterCategoryBudgetModel from '../QuarterCategoryBudget/model'
-import { QuarterCategoryBudgetAttributes } from '../QuarterCategoryBudget/types'
+import { validateCategoryBudgets } from '../QuarterCategoryBudget/utils'
 
 import { QuarterBudgetAttributes } from './types'
-
-function toProposalGrantCategory(category: string): ProposalGrantCategory {
-  switch (category) {
-    case snakeCase(ProposalGrantCategory.Platform):
-      return ProposalGrantCategory.Platform
-    case snakeCase(ProposalGrantCategory.InWorldContent):
-      return ProposalGrantCategory.InWorldContent
-    case snakeCase(ProposalGrantCategory.CoreUnit):
-      return ProposalGrantCategory.CoreUnit
-    case snakeCase(ProposalGrantCategory.Sponsorship):
-      return ProposalGrantCategory.Sponsorship
-    case snakeCase(ProposalGrantCategory.Accelerator):
-      return ProposalGrantCategory.Accelerator
-    case snakeCase(ProposalGrantCategory.SocialMediaContent):
-      return ProposalGrantCategory.SocialMediaContent
-    case snakeCase(ProposalGrantCategory.Documentation):
-      return ProposalGrantCategory.Documentation
-  }
-  throw new Error(`Attempted to parse an invalid ProposalGrantCategory ${category}`)
-}
-
-function getCategoryBudgetTotal(categoryPercentage: number, newQuarterBudget: QuarterBudgetAttributes) {
-  if (categoryPercentage < 0 || categoryPercentage > 100) {
-    throw new Error(`Invalid category percentage`)
-  }
-  return (categoryPercentage * newQuarterBudget.total) / 100
-}
+import { thereAreNoOverlappingBudgets } from './utils'
 
 export default class QuarterBudgetModel extends Model<QuarterBudgetAttributes> {
   static tableName = 'quarter_budgets'
@@ -44,7 +16,7 @@ export default class QuarterBudgetModel extends Model<QuarterBudgetAttributes> {
 
   static async createNewBudgets(transparencyBudgets: TransparencyBudget[]) {
     const existingBudgets = await this.find<QuarterBudgetAttributes>()
-    const newBudgets = []
+    const newBudgets: QuarterBudgetAttributes[] = []
     if (transparencyBudgets.length > existingBudgets.length) {
       const sortedBudgets = transparencyBudgets.sort((a, b) => Date.parse(b.start_date) - Date.parse(a.start_date))
       const now = new Date()
@@ -52,8 +24,8 @@ export default class QuarterBudgetModel extends Model<QuarterBudgetAttributes> {
         const startAt = new Date(transparencyBudget.start_date)
         if (!existingBudgets.some((existingBudget) => existingBudget.start_at === startAt)) {
           const finishAt = Time(transparencyBudget.start_date).add(4, 'months').toDate() //TODO a chequear
-          if (this.thereAreNoOverlappingBudgets(startAt, finishAt, existingBudgets)) {
-            this.validateCategoryBudgets(transparencyBudget)
+          if (thereAreNoOverlappingBudgets(startAt, finishAt, existingBudgets)) {
+            validateCategoryBudgets(transparencyBudget)
             const newQuarterBudget: QuarterBudgetAttributes = {
               id: uuid(),
               total: transparencyBudget.total,
@@ -62,14 +34,9 @@ export default class QuarterBudgetModel extends Model<QuarterBudgetAttributes> {
               created_at: now,
               updated_at: now,
             }
-            await this.create(newQuarterBudget)
-            try {
-              await this.createCategoryBudgets(newQuarterBudget, transparencyBudget)
-              newBudgets.push(newQuarterBudget)
-            } catch (e) {
-              await this.delete(newQuarterBudget)
-              throw e
-            }
+            await this.saveAndCreateCategoryBudgets(newQuarterBudget, transparencyBudget, (newBudget) =>
+              newBudgets.push(newBudget)
+            )
           } else {
             throw new Error(`There are overlapping budgets with: ${transparencyBudget}`)
           }
@@ -83,39 +50,18 @@ export default class QuarterBudgetModel extends Model<QuarterBudgetAttributes> {
     return [...existingBudgets, ...newBudgets]
   }
 
-  private static thereAreNoOverlappingBudgets(
-    startAt: Date,
-    finishAt: Date,
-    existingBudgets: QuarterBudgetAttributes[]
-  ) {
-    return !existingBudgets.some(
-      (existingBudget) =>
-        (existingBudget.start_at <= startAt && existingBudget.finish_at > startAt) ||
-        (existingBudget.start_at > startAt && existingBudget.start_at < finishAt)
-    )
-  }
-
-  private static validateCategoryBudgets(transparencyBudget: TransparencyBudget) {
-    if (Object.values(transparencyBudget.category_percentages).reduce((prev, next) => prev + next, 0) !== 100) {
-      throw new Error(`Categories percentages do not amount to 100 for budget: ${transparencyBudget}`)
-    }
-  }
-
-  private static async createCategoryBudgets(
+  private static async saveAndCreateCategoryBudgets(
     newQuarterBudget: QuarterBudgetAttributes,
-    transparencyBudget: TransparencyBudget
+    transparencyBudget: TransparencyBudget,
+    onSuccess: (newBudget: QuarterBudgetAttributes) => void
   ) {
-    for (const category of Object.keys(transparencyBudget.category_percentages)) {
-      const categoryPercentage = transparencyBudget.category_percentages[category]
-      const newQuarterCategoryBudget: QuarterCategoryBudgetAttributes = {
-        quarter_budget_id: newQuarterBudget.id,
-        category: toProposalGrantCategory(category),
-        total: getCategoryBudgetTotal(categoryPercentage, newQuarterBudget),
-        allocated: 0,
-        created_at: newQuarterBudget.created_at,
-        updated_at: newQuarterBudget.updated_at,
-      }
-      await QuarterCategoryBudgetModel.create(newQuarterCategoryBudget)
+    await this.create(newQuarterBudget)
+    try {
+      await QuarterCategoryBudgetModel.createCategoryBudgets(newQuarterBudget, transparencyBudget)
+      onSuccess(newQuarterBudget)
+    } catch (e) {
+      await this.delete(newQuarterBudget)
+      throw e
     }
   }
 }
