+++ src/entities/QuarterBudget/utils.ts
@@ -1,0 +1,13 @@
+import { QuarterBudgetAttributes } from './types'
+
+export function thereAreNoOverlappingBudgets(
+  startAt: Date,
+  finishAt: Date,
+  existingBudgets: QuarterBudgetAttributes[]
+) {
+  return !existingBudgets.some(
+    (existingBudget) =>
+      (existingBudget.start_at <= startAt && existingBudget.finish_at > startAt) ||
+      (existingBudget.start_at > startAt && existingBudget.start_at < finishAt)
+  )
+}
